requirements.txt
streamlit
langchain
ipykernel
python-dotenv 
langchain_groq 
langchain_core 
langserve 
streamlit
langchain_community
langchain_chroma 
langchain_huggingface 
sentence_transformers
bs4 
langchain-text-splitters
chromadb 
sentence_transformers
pypdf 
PyPDF2 
groq 
faiss-cpu 